+++ CHANGES.md
@@ -3,6 +3,7 @@ New release
 
 - add times tab to sprint view
 - tablesorter in time entries views i.e. /times/tickets/report
+- add bug sprint time in /scrum/sprint/show
 
 
 0.8 (23-04-2013)